fix: mediumテスト回帰を解消しサーバー初期化契約とアップロード検証を整合させる
server.js の起動契約を変更したことで medium の初期化テスト前提を崩し、
さらに Cookie 認証経由の media 登録テストでは画像シグネチャ検証と不整合なダミーファイルが原因で 400 となっていた。
回帰を防ぐため createEnv を分離してテスト可能性を確保しつつ server.js の自動起動挙動を復元し、
アップロード回帰テストは実際に許可される JPEG シグネチャを持つ入力へ修正して期待仕様に合わせた。
